+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Types of changes:
 
 ## [Unreleased]
 
+### Added
+- Tool to run `extract` and `build` steps on a single file (`eyex-extract-build`)
+- Tool to run built functions against individual files (`eyex-run-function`)
+- Problem list to items to be removed as part of the history section
+
 ## v20221108 - AMD Release 1
 
 ### Added